Change metrics API to use interfaces for Meter and BatchObserver
